HSEARCH-3619 Move bridge-related "mapping" classes to the "pojo.bridge.impl" package
